# docs/generated/api/swagger.yaml
definitions:
...
  contract.CreateTicketRequest:
    properties:
      category:
        type: string
      custom_fields:
        additionalProperties: true
        type: object
      customer_id:
        type: integer
      description:
        type: string
      priority:
        type: string
      session_id:
        type: string
      source:
        type: string
      tags:
        type: string
      title:
        type: string
    required:
    - customer_id
    - title
    type: object
...
  contract.TicketStats:
    properties:
      by_priority:
        items:
          $ref: '#/definitions/contract.PriorityCount'
        type: array
      by_status:
        items:
          $ref: '#/definitions/contract.StatusCount'
        type: array
      pending:
        type: integer
      resolved:
        type: integer
      today_created:
        type: integer
      total:
        type: integer
    type: object
...
  contract.UpdateTicketRequest:
    properties:
      agent_id:
        type: integer
      category:
        type: string
      custom_fields:
        additionalProperties: true
        type: object
      description:
        type: string
      due_date:
        type: string
      priority:
        type: string
      status:
        type: string
      tags:
        type: string
      title:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  handlers.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
...
  models.Customer:
    properties:
      company:
        type: string
      created_at:
        type: string
      id:
        type: integer
      industry:
        type: string
      notes:
        type: string
      priority:
        description: low, normal, high, urgent
        type: string
      source:
        description: web, referral, marketing
        type: string
      tags:
        description: 标签，逗号分隔
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      workspace_id:
        type: string
    type: object
  models.CustomerSatisfaction:
    properties:
      agent:
        $ref: '#/definitions/models.User'
      agent_id:
        type: integer
      category:
        description: service_quality, response_time, resolution_quality, overall
        type: string
      comment:
        type: string
      created_at:
        type: string
      customer:
        $ref: '#/definitions/models.Customer'
      customer_id:
        type: integer
      id:
        type: integer
      rating:
        description: 1-5星
        type: integer
      ticket:
        $ref: '#/definitions/models.Ticket'
      ticket_id:
        type: integer
    type: object
...
  models.SatisfactionSurvey:
    properties:
      agent_id:
        type: integer
      channel:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      expires_at:
        type: string
      id:
        type: integer
      satisfaction_id:
        type: integer
      sent_at:
        type: string
      status:
        description: queued, sent, completed, expired
        type: string
      survey_token:
        type: string
      ticket_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Ticket:
    properties:
      agent:
        $ref: '#/definitions/models.User'
      agent_id:
        type: integer
      attachments:
        items:
          $ref: '#/definitions/models.TicketFile'
        type: array
      category:
        description: technical, billing, general, complaint
        type: string
      closed_at:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.TicketComment'
        type: array
      created_at:
        type: string
      custom_field_values:
        items:
          $ref: '#/definitions/models.TicketCustomFieldValue'
        type: array
      customer:
        allOf:
        - $ref: '#/definitions/models.User'
        description: 关联关系
      customer_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      priority:
        description: low, normal, high, urgent
        type: string
      resolved_at:
        type: string
      session:
        $ref: '#/definitions/models.Session'
      session_id:
        type: string
      source:
        description: web, email, phone, chat
        type: string
      status:
        description: open, assigned, in_progress, resolved, closed
        type: string
      status_history:
        items:
          $ref: '#/definitions/models.TicketStatus'
        type: array
      tags:
        description: 标签，逗号分隔
        type: string
      tenant_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: string
    type: object
...
  services.SatisfactionCreateRequest:
    properties:
      agent_id:
        type: integer
      category:
        description: service_quality, response_time, resolution_quality, overall
        type: string
      comment:
        type: string
      customer_id:
        type: integer
      rating:
        maximum: 5
        minimum: 1
        type: integer
      ticket_id:
        type: integer
    required:
    - customer_id
    - rating
    - ticket_id
    type: object
  services.SatisfactionStat:
    properties:
      average_rating:
        type: number
      count:
        type: integer
    type: object
  services.SatisfactionStatsResponse:
    properties:
      average_rating:
        type: number
      category_stats:
        additionalProperties:
          $ref: '#/definitions/services.SatisfactionStat'
        type: object
      rating_distribution:
        additionalProperties:
          type: integer
        description: rating -> count
        type: object
      total_ratings:
        type: integer
      trend_data:
        items:
          $ref: '#/definitions/services.SatisfactionTrend'
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /api/satisfactions:
    get:
      description: 获取满意度评价列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页大小
        in: query
        name: page_size
        type: integer
      - description: 工单ID
        in: query
        name: ticket_id
        type: integer
      - description: 客户ID
        in: query
        name: customer_id
        type: integer
      - description: 客服ID
        in: query
        name: agent_id
        type: integer
      - collectionFormat: csv
        description: 评分筛选
        in: query
        items:
          type: integer
        name: rating
        type: array
      - collectionFormat: csv
        description: 分类筛选
        in: query
        items:
          type: string
        name: category
        type: array
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - default: created_at
        description: 排序字段
        in: query
        name: sort_by
        type: string
      - default: desc
        description: 排序顺序 (asc/desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CustomerSatisfaction'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取满意度评价列表
      tags:
      - 满意度评价
    post:
      consumes:
      - application/json
      description: 为已关闭的工单创建客户满意度评价
      parameters:
      - description: 满意度评价信息
        in: body
        name: satisfaction
        required: true
        schema:
          $ref: '#/definitions/services.SatisfactionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CustomerSatisfaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 创建满意度评价
      tags:
      - 满意度评价
  /api/satisfactions/{id}:
    delete:
      description: 删除指定的满意度评价（管理员功能）
      parameters:
      - description: 满意度评价ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: 删除成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 删除满意度评价
      tags:
      - 满意度评价
    get:
      description: 根据ID获取满意度评价的详细信息
      parameters:
      - description: 满意度评价ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerSatisfaction'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取满意度评价详情
      tags:
      - 满意度评价
    put:
      consumes:
      - application/json
      description: 更新满意度评价的评论（仅允许更新评论内容）
      parameters:
      - description: 满意度评价ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          properties:
            comment:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerSatisfaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 更新满意度评价
      tags:
      - 满意度评价
  /api/satisfactions/stats:
    get:
      description: 获取客户满意度的统计数据，包括平均评分、分布情况和趋势
      parameters:
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.SatisfactionStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取满意度统计
      tags:
      - 满意度评价
  /api/satisfactions/surveys:
    get:
      description: 查看调查发送状态、token 与完成情况
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 工单ID筛选
        in: query
        name: ticket_id
        type: integer
      - description: 客户ID筛选
        in: query
        name: customer_id
        type: integer
      - collectionFormat: multi
        description: 状态筛选
        in: query
        items:
          type: string
        name: status
        type: array
      - collectionFormat: multi
        description: 渠道筛选
        in: query
        items:
          type: string
        name: channel
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取 CSAT 调查列表
      tags:
      - 满意度评价
  /api/satisfactions/surveys/{id}/resend:
    post:
      description: 对指定调查重新生成链接与有效期
      parameters:
      - description: 调查ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SatisfactionSurvey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 重新发送 CSAT 调查
      tags:
      - 满意度评价
...
  /api/tickets:
    get:
      consumes:
      - application/json
      description: 获取工单列表，支持分页和过滤
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        type: integer
      - collectionFormat: csv
        description: 状态过滤
        in: query
        items:
          type: string
        name: status
        type: array
      - collectionFormat: csv
        description: 优先级过滤
        in: query
        items:
          type: string
        name: priority
        type: array
      - collectionFormat: csv
        description: 分类过滤
        in: query
        items:
          type: string
        name: category
        type: array
      - description: 客服ID过滤
        in: query
        name: agent_id
        type: integer
      - description: 客户ID过滤
        in: query
        name: customer_id
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - description: 排序字段
        in: query
        name: sort_by
        type: string
      - description: 排序方向
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Ticket'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取工单列表
      tags:
      - 工单
    post:
      consumes:
      - application/json
      description: 创建新的客服工单
      parameters:
      - description: 工单信息
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/contract.CreateTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Ticket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 创建工单
      tags:
      - 工单
  /api/tickets/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取工单的详细信息
      parameters:
      - description: 工单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取工单详情
      tags:
      - 工单
    put:
      consumes:
      - application/json
      description: 更新工单信息
      parameters:
      - description: 工单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: ticket
        required: true
        schema:
          $ref: '#/definitions/contract.UpdateTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 更新工单
      tags:
      - 工单
...
  /api/tickets/{ticket_id}/satisfaction:
    get:
      description: 根据工单ID获取对应的满意度评价
      parameters:
      - description: 工单ID
        in: path
        name: ticket_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerSatisfaction'
        "204":
          description: 未找到满意度评价
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 根据工单获取满意度评价
      tags:
      - 满意度评价
...
  /api/tickets/stats:
    get:
      consumes:
      - application/json
      description: 获取工单相关的统计数据
      parameters:
      - description: 客服ID，用于获取特定客服的统计
        in: query
        name: agent_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contract.TicketStats'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: 获取工单统计
      tags:
      - 工单
swagger: "2.0"